requirements.txt
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
filelock==3.13.1
Flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.43.1
fsspec==2023.10.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pipreqs==0.4.13
psutil==5.9.6
py-cpuinfo==9.0.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
simple-websocket==1.0.0
six==1.16.0
supervision==0.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.203
urllib3==2.0.7
Werkzeug==3.0.1
wsproto==1.2.0
yarg==0.1.9
zipp==3.17.0
